add changes

+++ nova_plugin/keypair.py
@@ -54,7 +54,7 @@ def create(nova_client, args, **kwargs):
     public_key = keypair.get('public_key')
 
     if use_external_resource(ctx, nova_client, KEYPAIR_OPENSTACK_TYPE):
-        if not pk_exists:
+        if private_key_path and not pk_exists:
             delete_runtime_properties(ctx, RUNTIME_PROPERTIES_KEYS)
             raise NonRecoverableError(
                 'Failed to use external keypair (node {0}): the public key {1}'
@@ -63,18 +63,24 @@ def create(nova_client, args, **kwargs):
                                       ctx.node.properties['resource_id'],
                                       private_key_path))
         return
-
-    # Raise error if the file existed and the public key is not provided
-    if pk_exists and not public_key:
+    elif private_key_path and pk_exists:
+        # Raise error if the file existed and the public key is not provided
         raise NonRecoverableError(
             "Can't create keypair - private key path already exists: {0}"
             .format(private_key_path))
+    elif not private_key_path and not public_key:
+        raise NonRecoverableError(
+            'One of the following must be provided: \n'
+            '- A public key, in the public_key under the keypair property. \n'
+            '- A private key path, in the private_key_path property. ('
+            'Not recommended.)'
+        )
 
     keypair = nova_client.keypairs.create(keypair['name'], public_key)
 
     set_openstack_runtime_properties(ctx, keypair, KEYPAIR_OPENSTACK_TYPE)
     # Write to private key if we do not provide public key
-    if not public_key:
+    if private_key_path and public_key:
         try:
             # write private key file
             _mkdir_p(os.path.dirname(private_key_path))
@@ -183,12 +189,20 @@ def creation_validation(nova_client, **kwargs):
 
 
 def _get_private_key_path():
-    return os.path.expanduser(ctx.node.properties[PRIVATE_KEY_PATH_PROP])
+    key_path = os.path.expanduser(ctx.node.properties.get(PRIVATE_KEY_PATH_PROP))
+    if key_path:
+        ctx.logger.warn(
+            'You have requested to save the private key to {key_path}. '
+            'You are strongly discouraged from saving private keys '
+            'to the file system. This feature is currently supported, '
+            'but will be removed in future releases.'.format(key_path=key_path)
+        )
+    return key_path
 
 
 def _delete_private_key_file():
     private_key_path = _get_private_key_path()
-    ctx.logger.debug('deleting private key file at {0}'.format(
+    ctx.logger.debug('Deleting private key file at {0}'.format(
         private_key_path))
     try:
         os.remove(private_key_path)
